services/semantic-adapter/requirements.txt
psycopg2-binary
pgvector
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
